requirements.txt
langchain
openai
youtube-transcript-api
pytube
faiss-cpu
sentence-transformers
python-dotenv